Data.TypeNums

Description

This module provides a unified interface for natural numbers, signed integers, and rationals at the type level, in a way fully compatible with existing code using type-level naturals.

Natural numbers are expressed as always, e.g. 5. Negative integers are written as Neg 3. Ratios are written as 3 :% 2.

There are some naming conflicts between this module and GHC.TypeLits, notably the comparison and arithmetic operators. This module reexports Nat, KnownNat, natVal and natVal' so you may import just this module and not GHC.TypeLits.

If you wish to use other functionality from GHC.TypeLits, this package also provides the module Data.TypeLits that includes (almost) full functionality from GHC.TypeLits, but with the conflicts resolving in this packages favour.

Type level numbers

Naturals

type Nat = Natural

A type synonym for Natural.

Previously, this was an opaque data type, but it was changed to a type synonym.

Since: base-4.16.0.0

class KnownNat (n :: Nat)

This class gives the integer associated with a type-level natural. There are instances of the class for every concrete literal: 0, 1, 2, etc.

Since: base-4.7.0.0

Minimal complete definition

natSing

natVal :: forall (n :: Nat) proxy. KnownNat n => proxy n -> Integer

Since: base-4.7.0.0

natVal' :: forall (n :: Nat). KnownNat n => Proxy# n -> Integer

Since: base-4.8.0.0

data SomeNat

This type represents unknown type-level natural numbers.

Since: base-4.10.0.0

Constructors

KnownNat n => SomeNat (Proxy n)

someNatVal :: Integer -> Maybe SomeNat

Convert an integer into an unknown type-level natural.

Since: base-4.7.0.0

Integers

data TInt

(Kind) An integer that may be negative.

Constructors

Pos Nat

Neg Nat

class KnownInt (n :: k)

This class gives the (value-level) integer associated with a type-level integer. There are instances of this class for every concrete natural: 0, 1, 2, etc. There are also instances of this class for every negated natural, such as Neg 1.

Minimal complete definition

intSing

intVal :: forall {k} (n :: k) proxy. KnownInt n => proxy n -> Integer

Get the value associated with a type-level integer

intVal' :: forall {k} (n :: k). KnownInt n => Proxy# n -> Integer

Get the value associated with a type-level integer. The difference between this function and intVal is that it takes a Proxy# parameter, which has zero runtime representation and so is entirely free.

data SomeInt

This type represents unknown type-level integers.

Since: 0.1.1

Constructors

KnownInt n => SomeInt (Proxy n)

someIntVal :: Integer -> SomeInt

Convert an integer into an unknown type-level integer.

Since: 0.1.1

Rationals

data Rat

Type constructor for a rational

Constructors

k :% Nat

class KnownRat (r :: k)

This class gives the (value-level) rational associated with a type-level rational. There are instances of this class for every combination of a concrete integer and concrete natural.

Minimal complete definition

ratSing

ratVal :: forall {k} proxy (r :: k). KnownRat r => proxy r -> Rational

Get the value associated with a type-level rational

ratVal' :: forall {k} (r :: k). KnownRat r => Proxy# r -> Rational

Get the value associated with a type-level rational. The difference between this function and ratVal is that it takes a Proxy# parameter, which has zero runtime representation and so is entirely free.

data SomeRat

This type represents unknown type-level integers.

Since: 0.1.1

Constructors

KnownRat r => SomeRat (Proxy r)

someRatVal :: Rational -> SomeRat

Convert a rational into an unknown type-level rational.

Since: 0.1.1

type family Simplify (x :: Rat) :: Rat where ...

Reduce a type-level rational into its canonical form

Since: 0.1.4

Equations

Simplify (x ':% y) = Quot ('Pos x) (GCD x y) ':% Quot y (GCD x y)

Simplify (x ':% y) = Quot x (GCD x y) ':% Quot y (GCD x y)

Type level numerical operations

Comparisons

type family (a :: k1) ==? (b :: k2) :: Bool where ... infix 4

Boolean type-level equals. Useful for e.g. If (x ==? 0)

Equations

(a :: Rat) ==? (b :: Rat) = Simplify a == Simplify b

(a :: Nat) ==? ('Pos a :: TInt) = 'True

('Pos a :: TInt) ==? (a :: Nat) = 'True

(a :: Nat) ==? (r :: Rat) = (a ':% 1) ==? r

(r :: Rat) ==? (a :: Nat) = r ==? (a ':% 1)

(a :: TInt) ==? (r :: Rat) = (a ':% 1) ==? r

(r :: Rat) ==? (a :: TInt) = r ==? (a ':% 1)

(a :: k) ==? (b :: k) = a == b

(_1 :: k1) ==? (_2 :: k2) = 'False

type (/=?) (a :: k1) (b :: k2) = Not (a ==? b) infix 4

Boolean type-level not-equals.

type (==) (a :: k1) (b :: k2) = (a ==? b) ~ 'True infix 4

Equality constraint, used as e.g. (x == 3) => _

type (/=) (a :: k1) (b :: k2) = (a ==? b) ~ 'False infix 4

Not-equal constraint

type family (a :: k1) <=? (b :: k2) :: Bool where ... infix 4

Boolean comparison of two type-level numbers

Equations

(a :: Nat) <=? (b :: Nat) = a <=? b

('Pos a :: TInt) <=? ('Pos b :: TInt) = a <=? b

('Neg _1 :: TInt) <=? ('Pos _2 :: TInt) = 'True

('Pos 0) <=? ('Neg 0) = 'True

('Pos _1 :: TInt) <=? ('Neg _2 :: TInt) = 'False

('Pos a :: TInt) <=? (b :: k2) = a <=? b

(a :: k1) <=? ('Pos b :: TInt) = a <=? b

0 <=? ('Neg 0) = 'True

(a :: Nat) <=? ('Neg _1 :: TInt) = 'False

('Neg a :: TInt) <=? (b :: Nat) = 'True

('Neg a :: TInt) <=? ('Neg b :: TInt) = b <=? a

(n ':% 0 :: Rat) <=? (_1 :: k2) = TypeError ('Text "Denominator must not equal 0") :: Bool

(_1 :: k1) <=? (n ':% 0 :: Rat) = TypeError ('Text "Denominator must not equal 0") :: Bool

(n1 ':% d1 :: Rat) <=? (n2 ':% d2 :: Rat) = (n1 * d2) <=? (n2 * d1)

(a :: k1) <=? (n ':% d :: Rat) = (a * d) <=? n

(n ':% d :: Rat) <=? (b :: k3) = n <=? (b * d)

(_1 :: k1) <=? (_2 :: k2) = TypeError ('Text "Incomparable") :: Bool

type (<=) (a :: k1) (b :: k2) = (a <=? b) ~ 'True infix 4

type (<) (a :: k1) (b :: k2) = (b <=? a) ~ 'False infix 4

type (>=) (a :: k1) (b :: k2) = (b <=? a) ~ 'True infix 4

type (>) (a :: k1) (b :: k2) = (a <=? b) ~ 'False infix 4

Arithmetic

Unary Operations

type family Abs (x :: k) :: k where ...

The absolute value of a type-level number

Since: 0.1.4

Equations

Abs (x :: Nat) = x

Abs ('Pos x :: TInt) = 'Pos x

Abs ('Neg x :: TInt) = 'Pos x

Abs (x ':% y :: Rat) = Simplify (Abs x ':% y)

type family Negate (x :: k) :: NegK k where ...

The result of negating a TInt

Since: 0.1.4

Equations

Negate (x :: Nat) = Negate ('Pos x)

Negate ('Pos 0) = 'Pos 0

Negate ('Neg 0) = 'Pos 0

Negate ('Pos x :: TInt) = 'Neg x

Negate ('Neg x :: TInt) = 'Pos x

Negate (x ':% y :: Rat) = Negate x ':% y

type family Recip (x :: k) :: Rat where ...

The reciprocal of a type-level number

Since: 0.1.4

Equations

Recip (x :: Nat) = 'Pos 1 ':% x

Recip ('Pos x :: TInt) = 'Pos 1 ':% x

Recip ('Neg x :: TInt) = 'Neg 1 ':% x

Recip ('Pos x ':% y :: Rat) = 'Pos y ':% x

Recip ('Neg x ':% y :: Rat) = 'Neg y ':% x

Recip (x ':% y :: Rat) = 'Pos y ':% x

type family Floor (x :: k) :: TInt where ...

Round a type-level number towards negative infinity

Since: 0.1.4

Equations

Floor (x :: Nat) = 'Pos x

Floor (x :: TInt) = x

Floor (x ':% 0 :: Rat) = TypeError ('Text "The denominator must not be 0") :: TInt

Floor (x ':% y :: Rat) = 'Pos (Div x y)

Floor (x ':% y :: Rat) = Div x y

type family Ceiling (x :: k) :: TInt where ...

Round a type-level number towards positive infinity

Since: 0.1.4

Equations

Ceiling (x :: Nat) = 'Pos x

Ceiling (x :: TInt) = x

Ceiling (x ':% 0 :: Rat) = TypeError ('Text "The denominator must not be 0") :: TInt

Ceiling (x ':% y :: Rat) = Add 1 (Floor (Sub x 1 ':% y))

type family Truncate (x :: k) :: TInt where ...

Round a type-level number towards zero

Since: 0.1.4

Equations

Truncate (x :: Nat) = 'Pos x

Truncate (x :: TInt) = x

Truncate (x ':% 0 :: Rat) = TypeError ('Text "The denominator must not be 0") :: TInt

Truncate (x ':% y :: Rat) = 'Pos (Quot x y)

Truncate (x ':% y :: Rat) = Quot x y

Binary Operations

type (+) (a :: k1) (b :: k2) = Add a b infixl 6

The sum of two type-level numbers

type (-) (a :: k1) (b :: k2) = Sub a b infixl 6

The difference of two type-level numbers

For the difference of two naturals a and b, a-b is also a natural, so only exists for a >= b.

type (*) (a :: k1) (b :: k2) = Mul a b infixl 7

The product of two type-level numbers.

Due to changes in GHC 8.6, using this operator infix and unqualified requires the NoStarIsType language extension to be active. See the GHC 8.6.x migration guide for details: https://ghc.haskell.org/trac/ghc/wiki/Migration/8.6

type (/) (a :: k1) (b :: k2) = RatDiv a b infixl 7

The ratio of two type-level numbers

type (^) (a :: k1) (b :: k2) = Exp a b infixr 8

A type-level number raised to an integer power. For Nat powers, the result kind is the same as the base. For TInt powers, the result kind is Rat.

Since: 0.1.4

type family DivMod (x :: k) (y :: Nat) :: (IntDivK k, IntDivK k) where ...

The quotient and remainder of a type-level integer and a natural number. For a negative dividend, the remainder part is positive such that x = q*y + r @since 0.1.4

Equations

DivMod (_1 :: k) 0 = TypeError ('Text "Divisor must not be 0") :: (IntDivK k, IntDivK k)

DivMod (x :: Nat) y = UnPos (DivModAux ('Pos x) y ('Pos 0)) :: (Nat, Nat)

DivMod ('Pos x :: TInt) y = DivModAux ('Pos x) y ('Pos 0)

DivMod ('Neg x :: TInt) y = DivModNegFixup (DivModAux ('Pos x) y ('Pos 0)) y

type family QuotRem (x :: k) (y :: Nat) :: (IntDivK k, IntDivK k) where ...

The quotient and remainder of a type-level integer and a natural number. For a negative dividend, the remainder part is negative such that x = q*y + r @since 0.1.4

Equations

QuotRem ('Neg x :: TInt) y = QuotRemFixup (DivMod ('Neg x) y) y

QuotRem (x :: k) y = DivMod x y

type family Div (x :: k) (y :: Nat) :: IntDivK k where ... infixl 7

The quotient of a type-level integer and a natural number.

Since: 0.1.4

Equations

Div (x :: k) y = Fst (DivMod x y)

type family Mod (x :: k) (y :: Nat) :: IntDivK k where ... infixl 7

The remainder of a type-level integer and a natural number For a negative number, behaves similarly to mod. @since 0.1.4

Equations

Mod (x :: k) y = Snd (DivMod x y)

type family Quot (x :: k) (y :: Nat) :: IntDivK k where ... infixl 7

The integer part of the result of dividing an integer by a natural number

Since: 0.1.4

Equations

Quot (x :: k) y = Fst (QuotRem x y)

type family Rem (x :: k) (y :: Nat) :: IntDivK k where ... infixl 7

The remainder of the result of dividing an integer by a natural number

Since: 0.1.4

Equations

Rem (x :: k) y = Snd (QuotRem x y)

type family GCD (x :: k1) (y :: k2) :: Nat where ...

The greatest common divisor of two type-level integers

Since: 0.1.4

Equations

GCD (x :: Nat) (y :: Nat) = GCDAux x y

GCD (x :: Nat) (y :: TInt) = GCDAux x (UnPos (Abs y) :: Nat)

GCD (x :: TInt) (y :: Nat) = GCDAux (UnPos (Abs x) :: Nat) y

GCD (x :: TInt) (y :: TInt) = GCDAux (UnPos (Abs x) :: Nat) (UnPos (Abs y) :: Nat)

type family IntLog (n :: Nat) (x :: k) :: TInt where ...

The floor of the logarithm of a type-level number NB. unlike Log2, Log n 0 here is a type error.

Since: 0.1.4

Equations

IntLog 0 (_1 :: k) = TypeError ('Text "Invalid IntLog base: 0") :: TInt

IntLog 1 (_1 :: k) = TypeError ('Text "Invalid IntLog base: 1") :: TInt

IntLog _1 0 = TypeError ('Text "IntLog n 0 is infinite") :: TInt

IntLog _1 ('Pos 0) = TypeError ('Text "IntLog n 0 is infinite") :: TInt

IntLog _1 ('Neg 0) = TypeError ('Text "IntLog n 0 is infinite") :: TInt

IntLog _1 (0 ':% _2 :: Rat) = TypeError ('Text "IntLog n 0 is infinite") :: TInt

IntLog _1 (_2 ':% 0 :: Rat) = TypeError ('Text "IntLog parameter has zero denominator") :: TInt

IntLog _1 ('Neg _2 :: TInt) = TypeError ('Text "IntLog of a negative does not exist") :: TInt

IntLog n (x :: Nat) = IntLog n ('Pos x)

IntLog n ('Pos x :: TInt) = IntLogAux n ('Pos x)

IntLog n (x :: Rat) = If (Floor x == 'Pos 0) (NegLogFudge n x (Negate (IntLogAux n (Floor (Recip x))))) (IntLog n (Floor x))

type Log2 (x :: k) = IntLog 2 x

The floor of the logarithm base 2 of a type-level number. Note that unlike Log2, this errors on Log2 0.

Since: 0.1.4