Data.TypeNums.Ints

Description

Type level integers can be used in the same way as type level naturals from GHC.TypeLits, for example 3. However, a minus sign is not recognised in this context, so a negative type-level integer is instead written as Neg n.

Construction

data TInt

(Kind) An integer that may be negative.

Constructors

Pos Nat

Neg Nat

Linking type and value level

class KnownInt (n :: k)

This class gives the (value-level) integer associated with a type-level integer. There are instances of this class for every concrete natural: 0, 1, 2, etc. There are also instances of this class for every negated natural, such as Neg 1.

Minimal complete definition

intSing

intVal :: forall {k} (n :: k) proxy. KnownInt n => proxy n -> Integer

Get the value associated with a type-level integer

intVal' :: forall {k} (n :: k). KnownInt n => Proxy# n -> Integer

Get the value associated with a type-level integer. The difference between this function and intVal is that it takes a Proxy# parameter, which has zero runtime representation and so is entirely free.

data SomeInt

This type represents unknown type-level integers.

Since: 0.1.1

Constructors

KnownInt n => SomeInt (Proxy n)

someIntVal :: Integer -> SomeInt

Convert an integer into an unknown type-level integer.

Since: 0.1.1