Data.TypeNums.Rats

Description

Type level rational numbers expressed as a ratio between a type-level integer and a type-level natural. For example Neg 3 :% 2.

See also: Data.TypeInts

Construction

data Rat

Type constructor for a rational

Constructors

k :% Nat

Linking type and value level

class KnownRat (r :: k)

This class gives the (value-level) rational associated with a type-level rational. There are instances of this class for every combination of a concrete integer and concrete natural.

Minimal complete definition

ratSing

ratVal :: forall {k} proxy (r :: k). KnownRat r => proxy r -> Rational

Get the value associated with a type-level rational

ratVal' :: forall {k} (r :: k). KnownRat r => Proxy# r -> Rational

Get the value associated with a type-level rational. The difference between this function and ratVal is that it takes a Proxy# parameter, which has zero runtime representation and so is entirely free.

data SomeRat

This type represents unknown type-level integers.

Since: 0.1.1

Constructors

KnownRat r => SomeRat (Proxy r)

someRatVal :: Rational -> SomeRat

Convert a rational into an unknown type-level rational.

Since: 0.1.1